@@ -9,16 +9,15 @@ android {
             storePassword 'droidwatchkey@123'
         }
     }
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.developers.paras.droidwatch"
-        minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 20
-        versionName "20.0"
+        minSdkVersion 21
+        compileSdk 35
+        targetSdkVersion 35
+        versionCode 21
+        versionName "21.0"
         multiDexEnabled true
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         debug {
@@ -43,24 +42,25 @@ android {
             signingConfig signingConfigs.config
         }
     }
+    namespace 'com.developers.paras.droidwatch'
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:customtabs:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.google.android.gms:play-services-ads:17.1.3'
-    implementation 'com.google.firebase:firebase-core:16.0.7'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    implementation 'androidx.browser:browser:1.8.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.gms:play-services-ads:23.3.0'
+    implementation 'com.google.firebase:firebase-core:21.1.1'
 //    implementation('com.crashlytics.sdk.android:crashlytics:2.9.8@aar') {
 //        transitive = true;
 //    }
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.13.2'
 }
 apply plugin: 'com.google.gms.google-services'
 //apply plugin: 'io.fabric'
