@@ -1,12 +1,13 @@
 package com.developers.paras.droidwatch;
 
 
-import android.app.FragmentTransaction;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -42,7 +43,7 @@ public class GuideToConnect extends Fragment {
             public void onClick(View v) {
 
                 if (fm!=null){
-                    final android.support.v4.app.FragmentTransaction ft = fm.beginTransaction();
+                    final FragmentTransaction ft = fm.beginTransaction();
                     ft.replace(R.id.device_list_layout, new DeviceListFragment());
                     ft.commit();
                 }
